+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/actions/RuleSetUtil.java
@@ -9,7 +9,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.regex.Pattern;
@@ -17,7 +16,6 @@ import javax.xml.stream.XMLOutputFactory;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
 
-import net.sourceforge.pmd.eclipse.ui.actions.internal.InternalRuleSetUtil;
 import net.sourceforge.pmd.lang.rule.Rule;
 import net.sourceforge.pmd.lang.rule.RuleSet;
 import net.sourceforge.pmd.lang.rule.RuleSetLoader;
@@ -67,37 +65,6 @@ public final class RuleSetUtil {
         return retainOnly(ruleSet, wantedRules);
     }
 
-    @Deprecated
-    public static RuleSet addExcludePatterns(RuleSet ruleSet, Collection<String> activeExclusionPatterns,
-            Collection<String> buildPathExcludePatterns) {
-        Set<String> newExcludePatterns = new HashSet<>(InternalRuleSetUtil.convert(ruleSet.getFileExclusions()));
-        newExcludePatterns.addAll(activeExclusionPatterns);
-        newExcludePatterns.addAll(buildPathExcludePatterns);
-        Set<String> newIncludePatterns = new HashSet<>(InternalRuleSetUtil.convert(ruleSet.getFileInclusions()));
-
-        return RuleSet.create(ruleSet.getName(),
-                ruleSet.getDescription(),
-                ruleSet.getFileName(),
-                InternalRuleSetUtil.convertStringPatterns(newExcludePatterns),
-                InternalRuleSetUtil.convertStringPatterns(newIncludePatterns),
-                ruleSet.getRules());
-    }
-
-    @Deprecated
-    public static RuleSet addIncludePatterns(RuleSet ruleSet, Collection<String> activeInclusionPatterns,
-            Collection<String> buildPathIncludePatterns) {
-        Set<String> newExcludePatterns = new HashSet<>(InternalRuleSetUtil.convert(ruleSet.getFileExclusions()));
-        Set<String> newIncludePatterns = new HashSet<>(InternalRuleSetUtil.convert(ruleSet.getFileInclusions()));
-        newIncludePatterns.addAll(activeInclusionPatterns);
-        newIncludePatterns.addAll(buildPathIncludePatterns);
-
-        return RuleSet.create(ruleSet.getName(), ruleSet.getDescription(),
-                ruleSet.getFileName(),
-                InternalRuleSetUtil.convertStringPatterns(newExcludePatterns),
-                InternalRuleSetUtil.convertStringPatterns(newIncludePatterns),
-                ruleSet.getRules());
-    }
-
     public static RuleSet newSingle(Rule rule) {
         return RuleSet.forSingleRule(rule);
     }
@@ -106,11 +73,6 @@ public final class RuleSetUtil {
         Set<Pattern> emptySet = Collections.emptySet();
         Set<Rule> emptyRules = Collections.emptySet();
         return RuleSet.create(name, description, null, emptySet, emptySet, emptyRules);
-    }
-
-    @Deprecated
-    public static RuleSet addRuleSetByReference(RuleSet ruleSet, RuleSet sourceRuleSet, boolean allRules) {
-        return addRuleSetByReference(ruleSet, sourceRuleSet);
     }
 
     public static RuleSet addRuleSetByReference(RuleSet ruleSet, RuleSet sourceRuleSet) {
@@ -162,24 +124,6 @@ public final class RuleSetUtil {
         return addRules(ruleSet, Collections.singleton(newRule));
     }
 
-    @Deprecated
-    public static RuleSet setExcludePatterns(RuleSet ruleSet, Collection<String> excludePatterns) {
-        return RuleSet.create(ruleSet.getName(), ruleSet.getDescription(),
-                ruleSet.getFileName(),
-                InternalRuleSetUtil.convertStringPatterns(excludePatterns),
-                ruleSet.getFileInclusions(),
-                ruleSet.getRules());
-    }
-
-    @Deprecated
-    public static RuleSet setIncludePatterns(RuleSet ruleSet, Collection<String> includePatterns) {
-        return RuleSet.create(ruleSet.getName(), ruleSet.getDescription(),
-                ruleSet.getFileName(),
-                ruleSet.getFileExclusions(),
-                InternalRuleSetUtil.convertStringPatterns(includePatterns),
-                ruleSet.getRules());
-    }
-
     public static RuleSet setNameDescription(RuleSet ruleSet, String name, String description) {
         return RuleSet.create(name, description, ruleSet.getFileName(),
                 ruleSet.getFileExclusions(),
@@ -193,16 +137,6 @@ public final class RuleSetUtil {
                 ruleSet.getFileExclusions(),
                 ruleSet.getFileInclusions(),
                 ruleSet.getRules());
-    }
-
-    @Deprecated
-    public static RuleSet addExcludePatterns(RuleSet rs, Collection<String> excludePatterns) {
-        return addExcludePatterns(rs, excludePatterns, new HashSet<>());
-    }
-
-    @Deprecated
-    public static RuleSet addIncludePatterns(RuleSet rs, Collection<String> includePatterns) {
-        return addIncludePatterns(rs, includePatterns, new HashSet<>());
     }
 
     public static RuleSet clearRules(RuleSet ruleSet) {